+++ plearn_learners_experimental/SurfaceTemplate/ScoreLayerVariable.h
@@ -79,7 +79,7 @@ public:
     int n_active_templates;
     int n_inactive_templates;
     bool normalize_by_n_features;
-    long seed_;
+    int32_t seed_;
     bool simple_mixture;
     VMat templates_source;
